Lots of changes for llvm14

+++ include/clasp/core/core.h
@@ -1020,8 +1020,8 @@ size_t lisp_lambdaListHandlerNumberOfSpecialVariables(LambdaListHandler_sp llh);
   DebugStream *lisp_debugLog();
 /*! Return a string representation of the object */
   string lisp_rep(T_sp obj);
-size_t lisp_badge(T_sp obj);
- size_t lisp_general_badge(General_sp obj);
+  uint32_t lisp_badge(T_sp obj);
+  uint32_t lisp_general_badge(General_sp obj);
   Symbol_sp lisp_internKeyword(const string &name);
   Symbol_sp lisp_intern(const string &name);
   Symbol_sp lisp_intern(const string &symbolName, const string &packageName);
